Refactor: move ThemeManager to :core:ui:theme:manager module
