+++ legacy/ui/base/src/main/java/com/fsck/k9/ui/base/K9Activity.kt
@@ -8,9 +8,9 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.core.app.ActivityCompat
 import androidx.lifecycle.asLiveData
-import app.k9mail.legacy.ui.theme.ThemeManager
 import com.fsck.k9.controller.push.PushController
 import java.util.Locale
+import net.thunderbird.core.ui.theme.manager.ThemeManager
 import org.koin.android.ext.android.inject
 
 abstract class K9Activity(private val themeType: ThemeType) : AppCompatActivity() {
