+++ legacy/ui/legacy/src/main/java/com/fsck/k9/activity/MessageCompose.java
@@ -106,8 +106,7 @@ import com.fsck.k9.message.SimpleMessageBuilder;
 import com.fsck.k9.message.SimpleMessageFormat;
 import app.k9mail.legacy.search.LocalSearch;
 import com.fsck.k9.ui.R;
-import com.fsck.k9.ui.base.K9Activity;
-import app.k9mail.legacy.ui.theme.ThemeManager;
+import com.fsck.k9.ui.base.K9Activity;;
 import com.fsck.k9.ui.compose.IntentData;
 import com.fsck.k9.ui.compose.IntentDataMapper;
 import com.fsck.k9.ui.compose.QuotedMessageMvpView;
@@ -117,6 +116,7 @@ import com.fsck.k9.ui.helper.SizeFormatter;
 import com.fsck.k9.ui.messagelist.DefaultFolderProvider;
 import com.google.android.material.dialog.MaterialAlertDialogBuilder;
 import com.google.android.material.textview.MaterialTextView;
+import net.thunderbird.core.ui.theme.manager.ThemeManager;
 import org.openintents.openpgp.OpenPgpApiManager;
 import org.openintents.openpgp.util.OpenPgpIntentStarter;
 import timber.log.Timber;
